# gh-91603: Speed up `UnionType` instantiation