--- a/Misc/NEWS.d/next/Core and Builtins/2022-04-23-22-08-34.gh-issue-91603.GcWEkK.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-04-23-22-08-34.gh-issue-91603.GcWEkK.rst
@@ -1,0 +1,2 @@
+Speed up :class:`types.UnionType` instantiation. Patch provided by Yurii
+Karabas.