--- a/Objects/unionobject.c
+++ b/Objects/unionobject.c
@@ -302,7 +302,7 @@ _Py_union_type_or(PyObject* self, PyObject* other)
 
     PyObject *tuple;
 
-    if (_PyUnion_Check(self) & _PyUnion_Check(other)) {
+    if (_PyUnion_Check(self) && _PyUnion_Check(other)) {
         tuple = merge_union_and_union(self, other);
     } else if (_PyUnion_Check(self)) {
         tuple = merge_union_and_obj(self, other);