--- a/Objects/unionobject.c
+++ b/Objects/unionobject.c
@@ -274,7 +274,8 @@ merge_obj_and_union(PyObject *left, PyObject *right) {
         }
         if (!is_duplicate) {
             Py_INCREF(arg);
-            PyTuple_SET_ITEM(tuple, pos++, arg);
+            PyTuple_SET_ITEM(tuple, pos, arg);
+            pos++;
         }
     }