#91603

### Summary
Remove `types.UnionType.__args__` recreation (`flatten_args` and `dedup_and_flatten_args `).
Use the fact that in the case when union with another `types.UnionType` that union has deduplicated normalized `__args__`.

As a result complexity from `O((M+N)^2)` was reduced to `O(M*N)`, where `M` - length of left union args, `N` - length of right union args. 

### Benchmarks:
| Benchmark                                  | this    | main                 |
|--------------------------------------------|:-------:|:--------------------:|
| int \| float                                | 67.5 ns | 121 ns: 1.80x slower |
| int \| (list \| float)                       | 153 ns  | 250 ns: 1.64x slower |
| (int \| float) \| list                       | 142 ns  | 238 ns: 1.67x slower |
| (int \| float) \| (list \| set)               | 253 ns  | 411 ns: 1.63x slower |
| (float \| set) \| int \| int \| (float \| list) | 450 ns  | 687 ns: 1.53x slower |
| float \| set \| int \| int \| float \| list     | 416 ns  | 750 ns: 1.80x slower |
| Geometric mean                             | (ref)   | 1.67x slower         |

For big unions:
| Benchmark                                                                                                                                          | this    | main                  |
|----------------------------------------------------------------------------------------------------------------------------------------------------|:-------:|:---------------------:|
| bool \| int \| float \| complex \| str \| bytes \| bytearray \| list \| set \| frozenset \| dict \| range \| property \| classmethod \| staticmethod \| Exception | 1.38 us | 4.08 us: 2.96x slower |