@uriyyo 2022-04-28
Closed because https://github.com/python/cpython/pull/91955 merged